[nv] M1 norm-epilogue absorption: single-round (half)((h*s)*w) per-load affine + fp16 store
- root cause (control-chain probe, /tmp/m1_norm_source_control.txt): the
  ordinary ffn-norm epilogue E_32_32_4_f14a5cc0 rounds ONCE at the very end
  ((half)((x*s)*w), weight upcast fp16->fp32); the old research kernel rounded
  twice (fp16(x*s), then fp16(*w)), used an fp16 pre-round of h, and computed
  scale via .sum() association -- 3 bitwise gaps vs control
- kernel q4k_g3_lanemap_gemv_w1w3_rms_affine_kernel: per-load affine is now the
  control epilogue's single fp16 RNE round; store_fp16=True adds the
  q4k_g3_lanemap_w1w3_rms_affine16_* name (mirrors the landed fused16 store so
  the ffn-activation cast folds and ffn_down sees the fp16 ABI)
- route q4k_gate_up_rms_affine_qualification_call: passes raw fp32 h (no fp16
  pre-round), scale is (h.float().square().mean(-1,keepdim=True)+eps).rsqrt()
  on the raw h so it renders the SAME r_16_256 reduce as control, fp16 typed
  output declaration (epilogue_absorption_admitted) under store_fp16
- model wiring: M1 lease checked BEFORE the M2b branch and ffn_down(z,
  normed_h=h) preserves the M2b in-kernel residual add
- tests: microgate + algebra unit tests updated to the single-round contract;
  NV microgate correctness now bitwise (max_abs 0.0)
- new harness extra/llm_research/decode/nv_epilogue_absorption_m1_ab.py:
  control = LANDED M2d candidate, candidate = +_rms_affine_gateup_norm_weight
  per block; smoke -> exact logits -> M1 census -> +50 us wall bracket

+++ test/unit/test_rmsnorm_scale_only_algebra.py
@@ -3,8 +3,10 @@ import numpy as np
 
 def _rms_affine(x, weight, eps=1e-6):
   scale=np.float32(1.0)/np.sqrt(np.mean(x.astype(np.float32)**2,dtype=np.float32)+np.float32(eps))
-  # This is the production fp16 round point: it cannot be moved after GEMV.
-  return ((x.astype(np.float32)*scale).astype(np.float16)*weight.astype(np.float16)).astype(np.float16), scale
+  # Production fp16 round point (E_32_32_4_f14a5cc0 contract): ONE round at the end of
+  # the fp32 affine chain -- (half)((x*s)*w), weight upcast fp16->fp32. It cannot be
+  # moved after GEMV (the consumer dots in fp32 on the fp16-rounded value).
+  return ((x.astype(np.float32)*scale)*weight.astype(np.float32)).astype(np.float16), scale
 
 
 def test_scale_only_consumer_must_apply_per_element_affine_before_q4_dot():
@@ -18,5 +20,5 @@ def test_scale_only_consumer_must_apply_per_element_affine_before_q4_dot():
   # nonuniform weight and an fp16 round point before the consumer.
   invalid=np.dot(row.astype(np.float32),x.astype(np.float32))*scale
   assert abs(float(reference-invalid)) > 1e-3
-  raw_affine=((x.astype(np.float32)*scale).astype(np.float16)*norm_weight).astype(np.float16)
+  raw_affine=((x.astype(np.float32)*scale)*norm_weight.astype(np.float32)).astype(np.float16)
   np.testing.assert_array_equal(raw_affine,normalized)
